requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
PyJWT==2.8.0
psycopg2-binary==2.9.9
# For testing
pytest==7.4.3
python-dotenv==1.0.0
customtkinter==5.2.1
requests==2.31.0
